+++ src/main/java/com/kidami/security/dto/lessonDTO/LessonUpdateDTO.java
@@ -11,15 +11,15 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class LessonUpdateDTO {
-    private Integer id;
+    private Long id;
     @Size(min = 2, max = 255, message = "Le nom doit contenir entre 2 et 255 caractères")
     private String name;
     @Size(max = 500, message = "La thumbnail ne doit pas dépasser 500 caractères")
     private String thumbnail;
     private String description;
-    private Integer courId;
+    private Long courId;
     private List<LessonVideoItemUpdateDTO> videos;
-    private List<Integer> videosToDelete;
+    private List<Long> videosToDelete;
 
     // Méthode pour vérifier si au moins un champ est rempli
     public boolean hasUpdates() {
