+++ src/main/java/com/kidami/security/repository/CategoryRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     Optional<Category> findByCategoryName(String categoryName);
     boolean existsByCategoryName(String categoryName);
     @Query("SELECT CASE WHEN COUNT(c) > 0 THEN true ELSE false END FROM Category c WHERE c.categoryName = :name AND c.categoryId != :id")
-    boolean existsByCategoryNameAndIdNot(@Param("name") String name, @Param("id") Integer id);
+    boolean existsByCategoryNameAndIdNot(@Param("name") String name, @Param("id") Long id);
 }
