+++ src/main/java/com/kidami/security/repository/LessonRepository.java
@@ -8,14 +8,14 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface LessonRepository extends JpaRepository<Lesson,Integer> {
+public interface LessonRepository extends JpaRepository<Lesson,Long> {
     Optional<Lesson> findByName(String name);
-    List<Lesson> findByCourId(Integer courId);
+    List<Lesson> findByCourId(Long courId);
 
-    Optional<Lesson> findById(Integer id);
-    boolean existsById(Integer id);
+    Optional<Lesson> findById(Long id);
+    boolean existsById(Long id);
     boolean existsByName(String name);
     @Query("SELECT CASE WHEN COUNT(c) > 0 THEN true ELSE false END FROM Lesson c WHERE c.name = :name AND c.id != :id")
-    boolean existsByNameAndIdNot(@Param("name") String name, @Param("id") Integer id);
+    boolean existsByNameAndIdNot(@Param("name") String name, @Param("id") Long id);
 
 }
